requirements.txt
# FastAPI 및 서버 기본
fastapi==0.104.1
mangum==0.17.0
python-dotenv==1.0.1
python-multipart==0.0.9

# OpenAI 및 LLM
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8

# 데이터 검증
pydantic==2.7.1

# HTTP 클라이언트
httpx==0.27.0
requests==2.31.0
